+++ examples/dummy-services/src/main/kotlin/de/gtarc/opaca/demo/reallab/WayfindingAgent.kt
@@ -25,18 +25,25 @@ val ROOMS = mapOf(
  * Since this service would have an effect on the real world, the dummy-version does not really
  * do anything, except finding the room id for a given room name.
  */
-class WayfindingAgent : AbstractContainerizedAgent(name="wayfinding-agent") {
-
+class WayfindingAgent : AbstractContainerizedAgent(
+    name = "wayfinding-agent",
+    description = "Agent for interacting with the ZEKI floor-based LED wayfinding system (dummy version)"
+) {
 
     override fun setupAgent() {
-        addAction("FindRoomById", mapOf("roomId" to Parameter("integer")), null) {
+        this.addAction("FindRoomById", "Show the way to the room with the given ID using the LED ground guidance system", mapOf(
+            "roomId" to Parameter("integer")
+        ), null) {
             val roomId = it["roomId"]!!.asInt()
-            startWayfinding(roomId)
+            this.startWayfinding(roomId)
         }
-        addAction("FindRoomByName", mapOf("roomName" to Parameter("string")), null) {
+
+        this.addAction("FindRoomByName", "Show the way to the room with the given name using the LED ground guidance system", mapOf(
+            "roomName" to Parameter("string")
+        ), null) {
             val roomName = it["roomName"]!!.asText()
             val roomId = this.getRoomIdFromHint(roomName)
-            startWayfinding(roomId)
+            this.startWayfinding(roomId)
         }
     }